Fireblaze

Use fireblaze to manage authentications.

Public fields

session

A valid Shiny session.

Active bindings

signed_in

Read the signed in user.

signed_up

Read the signed in user.

Methods

Public methods

Firebase$new()

Firebase$sign_out()

Firebase$get_signed_in()

Firebase$get_signed_up()

Firebase$is_signed_in()

Firebase$req_sign_in()

Firebase$set_language_code()

Firebase$delete_user()

Firebase$get_delete_user()

Firebase$clone()

Method new()

Usage

Firebase$new(
  persistence = c("none", "session", "local"),
  config_path = "firebase.rds",
  session = shiny::getDefaultReactiveDomain()
)

Arguments

persistence

How the auth should persit: none, the user has to sign in at every visit, session will only persist in current tab, local persist even when window is closed.

config_path

Path to the configuration file as created by create_config.

session

A valid shiny session.

Details

Initialise Fireblaze

Method sign_out()

Usage

Firebase$sign_out()

Details

Signs user out

Method get_signed_in()

Usage

Firebase$get_signed_in()

Details

Signed in user details triggered when auth states changes

Method get_signed_up()

Usage

Firebase$get_signed_up()

Details

Signed up user details triggered explicitely by user (e.g.: on click)

Method is_signed_in()

Usage

Firebase$is_signed_in()

Details

Check whether use is signed in

Method req_sign_in()

Usage

Firebase$req_sign_in()

Details

Makes output require user sign in

Method set_language_code()

Usage

Firebase$set_language_code(code)

Arguments

code

iso639-1 language code.

Details

Set language code for auth provider

Method delete_user()

Usage

Firebase$delete_user()

Details

Delete the user

Method get_delete_user()

Usage

Firebase$get_delete_user()

Details

Get result of user deletion

Method clone()

The objects of this class are cloneable with this method.

Usage

Firebase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.