OAuth Providers

Use OAuth provides such as Github or Facebook to allow users to conveniently sign in.

Super class

firebase::Firebase -> FirebaseOauthProviders

Methods

Public methods

FirebaseOauthProviders$set_provider()

FirebaseOauthProviders$launch()

FirebaseOauthProviders$clone()

Inherited methods

firebase::Firebase$clear()

firebase::Firebase$delete_user()

firebase::Firebase$get_access_token()

firebase::Firebase$get_delete_user()

firebase::Firebase$get_id_token()

firebase::Firebase$get_sign_out()

firebase::Firebase$get_signed_in()

firebase::Firebase$get_signed_up()

firebase::Firebase$initialize()

firebase::Firebase$is_signed_in()

firebase::Firebase$print()

firebase::Firebase$req_sign_in()

firebase::Firebase$req_sign_out()

firebase::Firebase$request_id_token()

firebase::Firebase$set_language_code()

firebase::Firebase$sign_out()

Method set_provider()

Usage

FirebaseOauthProviders$set_provider(provider)

Arguments

provider

The provider to user, e.g.: microsoft.com, yahoo.com or google.com.

Details

Define provider to use

Returns

self

Method launch()

Usage

FirebaseOauthProviders$launch(flow = c("popup", "redirect"))

Arguments

flow

Authentication flow, either popup or redirect.

Details

Launch sign in with Google.

Returns

self

Method clone()

The objects of this class are cloneable with this method.

Usage

FirebaseOauthProviders$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(shiny)
library(firebase)

ui <- fluidPage(
  useFirebase(),
  actionButton("signin", "Sign in with Microsoft", icon = icon("microsoft")),
  plotOutput("plot")
)

server <- function(input, output, session){
  f <- FirebaseOauthProviders$
    new()$
    set_provider("microsoft.com")

  observeEvent(input$signin, {
    f$launch()
  })

  output$plot <- renderPlot({
    f$req_sign_in()
    plot(cars)
  })
  
}

if (FALSE) shinyApp(ui, server)