Phone

Use mobile phone numbers to authenticate users.

Super classes

firebase::Firebase -> firebase::FirebaseAuth -> FirebasePhone

Methods

Public methods

FirebasePhone$new()

FirebasePhone$verify()

FirebasePhone$get_recaptcha()

FirebasePhone$confirm()

FirebasePhone$get_verification()

FirebasePhone$get_confirmation()

FirebasePhone$clone()

Inherited methods

firebase::Firebase$expose_app()

firebase::FirebaseAuth$clear()

firebase::FirebaseAuth$delete_user()

firebase::FirebaseAuth$expose_auth()

firebase::FirebaseAuth$get_access_token()

firebase::FirebaseAuth$get_delete_user()

firebase::FirebaseAuth$get_id_token()

firebase::FirebaseAuth$get_sign_out()

firebase::FirebaseAuth$get_signed_in()

firebase::FirebaseAuth$get_signed_up()

firebase::FirebaseAuth$is_signed_in()

firebase::FirebaseAuth$print()

firebase::FirebaseAuth$req_sign_in()

firebase::FirebaseAuth$req_sign_out()

firebase::FirebaseAuth$request_id_token()

firebase::FirebaseAuth$set_language_code()

firebase::FirebaseAuth$sign_out()

Method new()

Usage

FirebasePhone$new(
  persistence = c("session", "local", "memory"),
  config_path = "firebase.rds",
  language_code = NULL,
  session = shiny::getDefaultReactiveDomain()
)

Arguments

persistence

How the auth should persit: none, the user has to sign in at every visit, session will only persist in current tab, local persist even when window is closed.

config_path

Path to the configuration file as created by firebase_config.

language_code

Sets the language to use for the UI. Supported languages are listed here. Set to browser to use the default browser language of the user.

session

A valid shiny session.

Details

Initialiases Firebase Phone

Initialises the Firebase application client-side.

Method verify()

Usage

FirebasePhone$verify(number, id = NULL)

Arguments

number

Phone number of the user.

id

Id of the button that triggers verification. If this is NULL the user has to go through the recaptcha, if not NULL is invisible.

Details

Verify a phhone number

Method get_recaptcha()

Usage

FirebasePhone$get_recaptcha()

Details

Results from the recaptcha

Method confirm()

Usage

FirebasePhone$confirm(code)

Arguments

code

Confirmation code received by the user.

Details

Confirm a code

Method get_verification()

Usage

FirebasePhone$get_verification()

Details

Get Verification

Returns

A list with a boolean (success) indicating whether the operation was successful and a response contianing the response from Firebase.

Method get_confirmation()

Usage

FirebasePhone$get_confirmation()

Details

Get Confirmation

Returns

A list with a boolean (success) indicating whether the operation was successful and a response contianing the response from Firebase.

Method clone()

The objects of this class are cloneable with this method.

Usage

FirebasePhone$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.