Realtime Database

Access Firebase Realtime Database

Super class

firebase::Firebase

Methods

Public methods

RealtimeDatabase$new()

RealtimeDatabase$ref()

RealtimeDatabase$on_value()

RealtimeDatabase$set()

RealtimeDatabase$update()

RealtimeDatabase$delete()

RealtimeDatabase$clone()

Inherited methods

firebase::Firebase$expose_app()

firebase::Firebase$print()

Method new()

Usage

RealtimeDatabase$new(
  config_path = "firebase.rds",
  session = shiny::getDefaultReactiveDomain()
)

Arguments

config_path

Path to the configuration file as created by firebase_config.

session

A valid shiny session.

Details

Initialiases Firebase Storage

Initialises the Firebase Storage application client-side.

Method ref()

Usage

RealtimeDatabase$ref(path = NULL)

Arguments

path

Path to the database full URL to file.

Details

Reference

Creates a reference to a file or directory you want to operate on. Note that this reference persists, make sure you change it between operations.

Returns

Invisibly return the class instance.

Method on_value()

Usage

RealtimeDatabase$on_value(response, path = NULL)

Arguments

response

A boolean or character string. TRUE indicates that you want to capture the results of the file upload (e.g.: success or failed) with get_response method. FALSE indicates you do not want those results back. A character string is used as named of the response which then can be used in the get_response method.

path

Path to the database full URL to file.

Details

On Value

When path or ref sees an update it sends the new data to response.

Method set()

Usage

RealtimeDatabase$set(data, response = NULL, path = NULL)

Arguments

data

Dataset to upload.

response

A boolean or character string. TRUE indicates that you want to capture the results of the file upload (e.g.: success or failed) with get_response method. FALSE indicates you do not want those results back. A character string is used as named of the response which then can be used in the get_response method.

path

Path to the database full URL to file.

Details

Set Data

Pushes data to the database.

Method update()

Usage

RealtimeDatabase$update(data, response = NULL, path = NULL)

Arguments

data

Dataset to update.

response

A boolean or character string. TRUE indicates that you want to capture the results of the file upload (e.g.: success or failed) with get_response method. FALSE indicates you do not want those results back. A character string is used as named of the response which then can be used in the get_response method.

path

Path to the database full URL to file.

Details

Update Data

Update a record.

Method delete()

Usage

RealtimeDatabase$delete(response = NULL, path = NULL)

Arguments

response

A boolean or character string. TRUE indicates that you want to capture the results of the file upload (e.g.: success or failed) with get_response method. FALSE indicates you do not want those results back. A character string is used as named of the response which then can be used in the get_response method.

path

Path to the database full URL to file.

Details

Delete

Delete data to the database.

Method clone()

The objects of this class are cloneable with this method.

Usage

RealtimeDatabase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.