Requires Signout

Define UI element that requires no user to be signed in. This will hide them viusally if no user is signed in. Note that this is not secure as someone can easily change the CSS when visiting the page to reveal those elements.

reqSignout(...)

Arguments

...
Any valid tags.

See also