Analytics

Analytics

Analytics

Super class

firebase::Firebase -> Analytics

Methods

Public methods

Analytics$new()

Analytics$launch()

Analytics$enable()

Analytics$disable()

Analytics$log_event()

Analytics$set_user_properties()

Analytics$clone()

Inherited methods

firebase::Firebase$expose_app()

firebase::Firebase$print()

Method new()

Usage

Analytics$new(enable = TRUE)

Arguments

enable

Whether to internally enable analytics see launch method.

Details

Initialise

Initialise an analytics object.

Method launch()

Usage

Analytics$launch()

Details

Launch

Launch the analytics tracking. Note that analytics is not launched by the constructor in order to be able to enable or disable the tracking prior to the launch. This is because once Google Analytics is launched it cannot be disabled. If needed ask the user before running this method. The enabling and disabling of tracking provided by the package is only internal, e.g.: disabling tracking during a session will stop the log_event method from registering event but default Google Analytics will still be running.

Method enable()

Usage

Analytics$enable()

Details

Enable Tracking Internally enables tracking.

Method disable()

Usage

Analytics$disable()

Details

Disable Tracking Internally disables tracking: running methods from this instance of the class will not actually register with Google Analytics.

Method log_event()

Usage

Analytics$log_event(event, params = NULL)

Arguments

event

Event to log choose from this list of supported events.

params

Event parameters.

Details

Log Event

Log an event.

Method set_user_properties()

Usage

Analytics$set_user_properties(...)

Arguments

...

Named arguments defining the properties of the user.

Details

Set user properties

Method clone()

The objects of this class are cloneable with this method.

Usage

Analytics$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.