Prebuilt UI

Use fireblaze to manage authentications.

Super class

firebase::Firebase -> FirebaseUI

Public fields

tos_url

URL to the Terms of Service page.

privacy_policy_url

The URL to the Privacy Policy page.

Methods

Public methods

FirebaseUI$set_providers()

FirebaseUI$set_tos_url()

FirebaseUI$set_privacy_policy_url()

FirebaseUI$launch()

FirebaseUI$clone()

Inherited methods

firebase::Firebase$get_signed_in()

firebase::Firebase$get_signed_up()

firebase::Firebase$initialize()

firebase::Firebase$is_signed_in()

firebase::Firebase$req_sign_in()

firebase::Firebase$set_language_code()

firebase::Firebase$sign_out()

Method set_providers()

Usage

FirebaseUI$set_providers(
  google = FALSE,
  facebook = FALSE,
  twitter = FALSE,
  github = FALSE,
  email = FALSE,
  microsoft = FALSE,
  apple = FALSE,
  yahoo = FALSE,
  phone = FALSE,
  anonymous = FALSE
)

Arguments

google, facebook, twitter, github, email, microsoft, apple, yahoo, phone, anonymous

Set to TRUE the providers you want to use, at least one.

Details

Define signin and login providers.

Method set_tos_url()

Usage

FirebaseUI$set_tos_url(url)

Arguments

url

URL to use.

Details

Defines Tterms of Services URL

Method set_privacy_policy_url()

Usage

FirebaseUI$set_privacy_policy_url(url)

Arguments

url

URL to use.

Details

Defines Privacy Policy URL

Method launch()

Usage

FirebaseUI$launch(flow = c("popup", "redirect"), account_helper = FALSE)

Arguments

flow

The signin flow to use, popup or redirect.

account_helper

Wether to use accountchooser.com upon signing in or signing up with email, the user will be redirected to the accountchooser.com website and will be able to select one of their saved accounts. You can disable it by specifying the value below.

...

Any other option to pass to Firebase UI.

Details

Setup the signin form.

Method clone()

The objects of this class are cloneable with this method.

Usage

FirebaseUI$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.